requirements.txt
streamlit
crewai
crewai-tools
pinecone
groq
langchain-groq
sentence-transformers
beautifulsoup4
requests
pyyaml
pandas
matplotlib
plotly
python-dotenv
pytest
serpapi-python
